requirements.txt
tweepy==4.14.0
python-telegram-bot==21.5
APScheduler==3.10.4
aiohttp==3.9.5
httpx==0.27.0
